Add Node.js native addon for C interpreter (#2)

+++ src/native/interpret.h
@@ -2,7 +2,7 @@
 typedef struct {
   int sp;
   int ip;
-  int* code;
+  uint8_t* code;
   int code_length;
   int* stack;
   int stack_length;
